Added flipping function
I realized that current fish image from camera flipped 180degree of x axis. so I added flipping function to the scanner module. of cause we can solve this problem to install camera with 180 apposite from current. but it is better to solve this problem by a code. because we don't know yet about how camera would be installed to new scanner rig.